requirements.txt
aiohttp==3.9.5
fastapi==0.111.0
pydantic-settings==2.3.3
sqlalchemy==2.0.30
bcrypt==4.1.3
aiosqlite==0.20.0
pyjwt==2.8.0